feat: Display daily puzzle number on Sliding Puzzle
Adds the functionality to display the current day's puzzle number on the Sliding Puzzle game when in 'daily15' mode. This helps users identify which day's puzzle they are playing.

The `getPuzzleNumber` utility function was also updated to use a more robust date calculation method and changed the default `startDateStr` to '2025-09-19' to align with the intended puzzle numbering.